Fix for concurrent action calls
If multiple actions are called to the same endpoint in rapid succession, all the callbacks are pushed onto the same pubsub object. When the first fetch returns, all of the callbacks are called with that data, and subsequent fetch returns call no callbacks. Moving the pubsub instantiation into the individual calls to the api solves this issue.